requirements.txt
blessed==1.20.0
boto3==1.35.93
botocore==1.35.93
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
dnspython==2.7.0
editor==1.6.6
inquirer==3.4.0
jmespath==1.0.1
mysql-connector-python==9.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pymongo==4.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
readchar==4.2.1
runs==1.2.2
s3transfer==0.10.4
six==1.17.0
termcolor==2.3.0
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
xmod==1.8.1
yaspin==3.1.0